tool: Address MinGW compile warning
Warning was:
format ‘%lu’ expects argument of type ‘long unsigned int’, but argument 4 has type ‘int’ [-Wformat=]

+++ tool/uriparse.c
@@ -76,7 +76,8 @@ int main(int argc, char *argv[]) {
 						: (state.errorCode == URI_ERROR_MALLOC)
 							? "not enough memory"
 							: "liburiparser bug (please report)",
-					state.errorPos, state.errorPos - argv[i]);
+					state.errorPos,
+					(long unsigned int)(state.errorPos - argv[i]));
 			retval = EXIT_FAILURE;
 		} else {
 			if (uri.scheme.first) {
